requirements.txt
pyserial==3.5
gpiozero==1.6
evdev==1.7
